+++ src/commands/database.js
@@ -5,20 +5,19 @@ const { getBackups } = require('@clevercloud/client/cjs/api/v2/backups.js');
 const { println } = require('../logger.js');
 const formatTable = require('../format-table')();
 const superagent = require('superagent');
-const { getSummary } = require('@clevercloud/client/cjs/api/v2/user.js');
 const fs = require('fs');
+const { findOwnerId } = require('../models/addon.js');
+const { resolveRealId } = require('../models/ids-resolver.js');
 
 async function listBackups (params) {
 
-  const [databaseId] = params.args;
   const { org } = params.options;
+  const [addonIdOrRealId] = params.args;
 
-  const ownerId = await resolveOwnerId(org, databaseId);
-  if (ownerId == null) {
-    throw new Error('organisation ID is mandatory');
-  }
+  const addonId = await resolveRealId(addonIdOrRealId);
+  const ownerId = await findOwnerId(org, addonId);
 
-  const backups = await getBackups({ ownerId, ref: databaseId }).then(sendToApi);
+  const backups = await getBackups({ ownerId, ref: addonId }).then(sendToApi);
 
   if (backups.length === 0) {
     println('There are no backups yet');
@@ -47,15 +46,13 @@ async function listBackups (params) {
 
 async function downloadBackups (params) {
 
-  const [databaseId, backupId] = params.args;
   const { org, output } = params.options;
+  const [addonIdOrRealId, backupId] = params.args;
 
-  const ownerId = await resolveOwnerId(org, databaseId);
-  if (ownerId == null) {
-    throw new Error('organisation ID is mandatory');
-  }
+  const addonId = await resolveRealId(addonIdOrRealId);
+  const ownerId = await findOwnerId(org, addonId);
 
-  const backups = await getBackups({ ownerId, ref: databaseId }).then(sendToApi);
+  const backups = await getBackups({ ownerId, ref: addonId }).then(sendToApi);
   const backup = backups.find((backup) => backup.backup_id === backupId);
 
   if (backup == null) {
@@ -72,35 +69,6 @@ async function downloadBackups (params) {
   }
 
   process.stdout.write(res.body);
-}
-
-/**
- * Try to get an ownerId from the API
- * @param {*} org cliparse param's option for orga
- * @param {String} databaseId the resource which belong to the owner we are looking for
- * @returns the owner ID (or null if cannot be found)
- */
-async function resolveOwnerId (org, databaseId) {
-
-  if (org != null && org.orga_name != null) {
-    return org.orga_name;
-  }
-
-  const summary = await getSummary().then(sendToApi);
-
-  const userHasAddon = summary.user.addons.some((addon) => addon.realId === databaseId);
-  if (userHasAddon) {
-    return summary.user.id;
-  }
-
-  for (const orga of summary.organisations) {
-    const orgaHasAddon = orga.addons.some((addon) => addon.realId === databaseId);
-    if (orgaHasAddon) {
-      return orga.id;
-    }
-  }
-
-  return null;
 }
 
 module.exports = { listBackups, downloadBackups };