+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # clever-tools changelog
 
 ## Unrelease (????-??-??)
+
 * skip preorder step on addon creation
+* add `clever addon env` command
+* improve implicit ID params (owner and add-on ID / real ID) for `clever database`
 
 ## 2.10.1 (2023-02-20)